refactor: split analysis/runner.py into focused modules
